requirements.txt
flask
python-dotenv
watchdog
b2sdk
humanize
